Move particleset to core
