+++ parcels/particlefile.py
@@ -20,7 +20,7 @@ from parcels.utils._helpers import timedelta_to_float
 if TYPE_CHECKING:
     from parcels._core.utils.time import TimeInterval
     from parcels.particle import Variable
-    from parcels.particleset import ParticleSet
+    from parcels._core.particleset import ParticleSet
 
 __all__ = ["ParticleFile"]
