+++ tests/test_advection.py
@@ -25,7 +25,7 @@ from parcels.kernels import (
 )
 from parcels.particle import Particle, Variable
 from parcels.particlefile import ParticleFile
-from parcels.particleset import ParticleSet
+from parcels import ParticleSet
 from parcels.statuscodes import StatusCode
 from parcels import XGrid
 from tests.utils import round_and_hash_float_array
