+++ tests/test_interpolation.py
@@ -18,7 +18,7 @@ from parcels.interpolators import (
 from parcels.kernels import AdvectionRK4_3D
 from parcels.particle import Particle, Variable
 from parcels.particlefile import ParticleFile
-from parcels.particleset import ParticleSet
+from parcels import ParticleSet
 from parcels.statuscodes import StatusCode
 from parcels import UxGrid
 from parcels import XGrid
